feat: add background flow routes and integrate with Fastify server
