+++ backend/src/routes/backgrounds.ts
@@ -5,9 +5,6 @@ import {
   CreateBackgroundRequest, 
   UpdateBackgroundRequest,
   BackgroundListQuery,
-  BackgroundFlow,
-  BackgroundStep,
-  CreateBackgroundFlowRequest,
   ApiResponse 
 } from '../types/api'
 
@@ -358,175 +355,6 @@ export default async function backgroundsRoutes(fastify: FastifyInstance) {
 
       await fastify.prisma.background.delete({
         where: { id: backgroundId }
-      })
-
-      return reply.status(204).send()
-    } catch (error) {
-      fastify.log.error(error)
-      return reply.status(500).send({
-        success: false,
-        error: 'Internal server error'
-      } as ApiResponse)
-    }
-  })
-
-
-  // GET /api/v1/writers/:writerId/backgrounds/flow - 배경 플로우 전체 조회
-  fastify.get<{ Params: { writerId: string } }>('/api/v1/writers/:writerId/backgrounds/flow', {
-    schema: {
-      params: {
-        type: 'object',
-        required: ['writerId'],
-        properties: {
-          writerId: { type: 'string' }
-        }
-      }
-    }
-  }, async (request, reply) => {
-    try {
-      const { writerId } = request.params
-
-      const flows = await fastify.prisma.backgroundFlow.findMany({
-        where: { writerId },
-        include: {
-          backgroundSteps: {
-            include: {
-              background: true
-            },
-            orderBy: { orderKey: 'asc' }
-          }
-        },
-        orderBy: { createdAt: 'desc' }
-      })
-
-      const response: BackgroundFlow[] = flows.map(flow => ({
-        flowId: flow.id,
-        writerId: flow.writerId,
-        version: flow.version,
-        createdAt: flow.createdAt.toISOString(),
-        steps: flow.backgroundSteps.map(step => ({
-          stepId: step.id,
-          backgroundId: step.backgroundId,
-          orderKey: step.orderKey || 0
-        }))
-      }))
-
-      return reply.send({
-        success: true,
-        data: response
-      } as ApiResponse)
-    } catch (error) {
-      fastify.log.error(error)
-      return reply.status(500).send({
-        success: false,
-        error: 'Internal server error'
-      } as ApiResponse)
-    }
-  })
-
-  // POST /api/v1/writers/:writerId/backgrounds/flow - 배경 플로우 생성
-  fastify.post<{ 
-    Params: { writerId: string },
-    Body: CreateBackgroundFlowRequest 
-  }>('/api/v1/writers/:writerId/backgrounds/flow', {
-    schema: {
-      params: {
-        type: 'object',
-        required: ['writerId'],
-        properties: {
-          writerId: { type: 'string' }
-        }
-      },
-      body: {
-        type: 'object',
-        required: ['steps'],
-        properties: {
-          steps: {
-            type: 'array',
-            items: {
-              type: 'object',
-              required: ['backgroundId', 'orderKey'],
-              properties: {
-                backgroundId: { type: 'string' },
-                orderKey: { type: 'number' }
-              }
-            }
-          }
-        }
-      }
-    }
-  }, async (request, reply) => {
-    try {
-      const { writerId } = request.params
-      const { steps } = request.body
-
-      // TODO: Check if user has permission to create flow for this writer
-
-      const flow = await fastify.prisma.backgroundFlow.create({
-        data: {
-          writerId,
-          version: '1.0',
-          backgroundSteps: {
-            create: steps.map(step => ({
-              backgroundId: step.backgroundId,
-              orderKey: step.orderKey
-            }))
-          }
-        },
-        include: {
-          backgroundSteps: {
-            include: {
-              background: true
-            },
-            orderBy: { orderKey: 'asc' }
-          }
-        }
-      })
-
-      const response: BackgroundFlow = {
-        flowId: flow.id,
-        writerId: flow.writerId,
-        version: flow.version || '',
-        createdAt: flow.createdAt.toISOString(),
-        steps: flow.backgroundSteps.map(step => ({
-          stepId: step.id,
-          backgroundId: step.backgroundId,
-          orderKey: step.orderKey || 0
-        }))
-      }
-
-      return reply.status(201).send({
-        success: true,
-        data: response
-      } as ApiResponse)
-    } catch (error) {
-      fastify.log.error(error)
-      return reply.status(400).send({
-        success: false,
-        error: 'Invalid request data'
-      } as ApiResponse)
-    }
-  })
-
-  // DELETE /api/v1/writers/:writerId/backgrounds/flow - 플로우 삭제
-  fastify.delete<{ Params: { writerId: string } }>('/api/v1/writers/:writerId/backgrounds/flow', {
-    schema: {
-      params: {
-        type: 'object',
-        required: ['writerId'],
-        properties: {
-          writerId: { type: 'string' }
-        }
-      }
-    }
-  }, async (request, reply) => {
-    try {
-      const { writerId } = request.params
-
-      // TODO: Check if user has permission to delete flow for this writer
-
-      await fastify.prisma.backgroundFlow.deleteMany({
-        where: { writerId }
       })
 
       return reply.status(204).send()
